Move some test functions into a testing support module.
